Se separa la logica de pagos a un repo nuevo
