+++ sprint 2/BackendGeems/BackendGeems/Application/IGEEMSRepo.cs
@@ -1,19 +1,21 @@
 ﻿using BackendGeems.Domain;
+using static BackendGeems.Infraestructure.GEEMSRepo;
 namespace BackendGeems.Application
 {
     public interface IGEEMSRepo
     {
-        bool calcularPago(string fechaInicio, string fechaFinal);
-        List<Pago> ObtenerPagos(DateTime fechaInicio, DateTime fechaFinal);
-        int ObtenerSalarioBruto(Guid idEmpleado, DateTime fechaInicio, DateTime fechaFinal);
-        void GenerarPagoEmpleado(Guid idEmpleado, Guid idPlanilla, DateTime fechaInicio, DateTime fechaFinal);
-        void InsertDeduccion(Guid idPago, string tipo, Guid? idBeneficio, int monto);
+
+
+
         Registro GetRegister(Guid Id);
         void EditRegister(Registro editing, Guid oldId);
-        int CalcularImpuestoRenta(int salarioBruto);
+
         List<Registro> ObtenerRegistros(Guid idEmpleado);
         void InsertRegister(Registro registro);
         public int GetMonthHours(Guid idEmpleado, DateTime fecha);
+        public List<Empleado> ObtenerEmpleadosPorEmpresa(string nombreEmpresa);
+        public List<PlanillaDTO> ObtenerPlanillasPorEmpresa(string nombreEmpresa);
+
     }
 
 }
