+++ sprint 2/BackendGeems/BackendGeems/Program.cs
@@ -22,6 +22,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
 builder.Services.AddScoped<IGEEMSRepo, GEEMSRepo>();
+builder.Services.AddScoped<IGEEMSPagoRepo, GEEMSPagoRepo>();
 builder.Services.AddScoped<IQueryPago, QueryPago>();
 builder.Services.AddScoped<ISalarioBruto, SalarioBruto>();
 builder.Services.AddScoped<IGenerarPago, GenerarPago>();
